Might have fixed a bug in pa1

+++ arch/x86/kernel/csci3753_mult.c
@@ -5,8 +5,8 @@
 SYSCALL_DEFINE3(csci3753_mult, int, number1, int, number2, long *, result)
 {
   printk(KERN_ALERT "Multiplying '%d' with '%d'!\n", number1, number2);
-  int tmp_result = number1 * number2;
-  if(copy_to_user(result, &tmp_result, sizeof(int)))
+  long tmp_result = number1 * number2;
+  if(copy_to_user(result, &tmp_result, sizeof(long)))
   {
     printk(KERN_ALERT "Failed to copy data to user space\n");
     return -EFAULT;
